requirements.txt
Django~=5.0.3
djangorestframework~=3.15.1
pip~=24.0
attrs~=23.2.0
pillow~=10.3.0
asgiref~=3.8.1
sqlparse~=0.4.4
psycopg2~=2.9.9
config~=0.5.1
inflection~=0.5.1
openapi~=1.1.0